+++ transmart-rest-api/grails-app/services/org/transmartproject/rest/dataExport/ExportAsyncJobService.groovy
@@ -2,8 +2,13 @@ package org.transmartproject.rest.dataExport
 
 import grails.transaction.Transactional
 import org.grails.web.json.JSONObject
+import org.quartz.JobBuilder
+import org.quartz.JobDataMap
+import org.quartz.JobDetail
 import org.quartz.JobKey
 import org.quartz.Scheduler
+import org.quartz.TriggerBuilder
+import org.transmartproject.core.exceptions.InvalidRequestException
 import org.transmartproject.db.job.AsyncJobCoreDb
 import org.transmartproject.db.multidimquery.query.InvalidQueryException
 import org.transmartproject.db.user.User
@@ -28,6 +33,10 @@ class ExportAsyncJobService {
         }
     }
 
+    Boolean isJobNameUniqueForUser(String name, User user) {
+        getJobByNameAndUser(name, user) == null
+    }
+
     AsyncJobCoreDb getJobByNameAndUser(String jobName, User user) {
         def jobs = AsyncJobCoreDb.findAllByJobNameAndUserId(jobName, user.username)
         if(jobs.size() > 1) {
@@ -37,6 +46,7 @@ class ExportAsyncJobService {
     }
 
     AsyncJobCoreDb createNewJob(User user, String jobName = null) {
+
         def jobStatus = JobStatus.CREATED
 
         def newJob = new AsyncJobCoreDb(lastRunOn: new Date())
@@ -50,6 +60,8 @@ class ExportAsyncJobService {
             newJob.jobName = newJob.id.toString()
             newJob.save()
         }
+
+        log.debug("Sending ${newJob.jobName} back to the client")
 
         newJob
     }
@@ -94,4 +106,53 @@ class ExportAsyncJobService {
         }
         isJobCancelled
     }
+
+    private static Map createJobDataMap(List<Long> ids, String typeOfSet, List types, User user, Long jobId, String jobName) {
+        [
+                user                 : user,
+                jobName              : jobName,
+                jobId                : jobId,
+                ids                  : ids,
+                typeOfSet            : typeOfSet,
+                dataTypeAndFormatList: types
+        ]
+    }
+
+    private AsyncJobCoreDb executeExportJob(Map dataMap) {
+
+        def job
+        def jobDataMap = new JobDataMap(dataMap)
+        JobDetail jobDetail = JobBuilder.newJob(ExportJobExecutor.class)
+                .withIdentity(dataMap.jobId.toString(), 'DataExport')
+                .setJobData(jobDataMap)
+                .build()
+        def randomDelay = Math.random()*10 as int
+        def startTime = new Date(new Date().time + randomDelay)
+        def trigger = TriggerBuilder.newTrigger()
+                .startAt(startTime)
+                .withIdentity('DataExport')
+                .build()
+
+        if(!quartzScheduler) {
+            job = updateStatus(dataMap.jobId, JobStatus.ERROR, null,
+                    "No qualifying bean found for dependency 'org.quartz.Scheduler'")
+        } else {
+            job = updateStatus(dataMap.jobId, JobStatus.STARTED)
+            quartzScheduler.scheduleJob(jobDetail, trigger)
+        }
+
+        return job
+    }
+
+    def exportData(List<Long> ids, String typeOfSet, List types, User user, Long jobId) {
+        def job = getJobById(jobId)
+        if (job.jobStatus != JobStatus.CREATED.value) {
+            throw new InvalidRequestException("Job with id $jobId has invalid status. " +
+                    "Expected: $JobStatus.CREATED.value, actual: $job.jobStatus")
+        }
+
+        def dataMap = createJobDataMap(ids, typeOfSet, types, user, jobId, job.jobName)
+        executeExportJob(dataMap)
+    }
+
 }
